Refactor component name and add sign up page (non-functional)
